Allow save+run after downloading
The "download" buttons mutate the form action permanently, which means
that subsequent clicks on "save/save and run" then just trigger the file
to download again.

This patch removes any action from the form, which makes it post back to
the current URL and allows save/save+run to complete again.