fix: poll schema view until schema is returned from cache
In MULTIUSER_DEPLOYMENT mode (i.e not dataworkspace), the schema info is generated asynchronously and cached when a cache miss happens in the show schema view.

However, the cache miss results in the schema pane displaying the message "Schema is building" and the page gets stuck in that state. It's only when you press hide and show schema again that you see the schema.

This commit splits out the logic into two separate views; one which renders the schema pane template and one which just returns the schema html to be inserted into the div. If there is a cache miss, the iframe polls the view until the schema html is returned.