feature: scheduled schema build
In a multiuser deployment, the schema building is triggered when the user clicks on show schema.
When there are many tables this can take a while. We change the triggering to a scheduled task to populate
the cache every 8 minutes and the cache will expire every 10 minutes so the schema is quite up to date and
is always availble for quick retrieval from the cache.

In data workspace we assume the schema doesn't change during a user session as the user permissions cannot change.
So the cache is set on the start of the session and does not expire.